Remove deprecated ZipSlices (didn't have desired performance)
This is superseded by .zip()'s specialization in libstd, which is the
way forward.
